# [pkg/util/log] Scrub 'pwd' and certificates when sending flare