### What does this PR do?

`pwd` has been added to the list of password keys to scrub if encountered.
To remove private certificates / keys, multiline replacers have been introduced in the credentials scrubbing phase (since certificates are usually written on multiple lines).
Searching for certificates / keys is done by a regex which follows [RFC 7468](https://tools.ietf.org/html/rfc7468)'s ABNF as closely as possible. However, the lack of backreferences makes it impossible to check in the regex that the same label is used in the `BEGIN` and `END` parts.
Tests have been added for these two cases.

### Motivation

The `pwd` and `cert_key` fields of [this integration config file](https://github.com/DataDog/integrations-core/blob/c603898bbcb7a9d2801322b8353d729e54dcf6a1/cisco_aci/datadog_checks/cisco_aci/data/conf.yaml.example) weren't properly scrubbed when sending a flare.