@olivielpeau 2019-05-22
haven't reviewed the changes by themselves, a few notes/pointers though :) :

* `CredentialsCleanerBytes` is used by the logger on every log message that's logged, so is performance sensitive. Could you make sure to add the results (before/after your changes) of the benchmarks provided here? https://github.com/DataDog/datadog-agent/blob/master/pkg/util/log/log_bench_test.go
* the `cisco_aci` integration file also has a `cert_key_password` field, is it already being cleaned up as expected? (might be worth adding a test case in any case)