@KSerrania 2019-05-22
@olivielpeau I added the benchmarks and the `cert_key_password` test case.