@KSerrania 2019-05-28
New benchmarks (with `Hint` added)

Before change:
```
goos: darwin
goarch: amd64
pkg: github.com/DataDog/datadog-agent/pkg/util/log
BenchmarkLogVanilla-8                1000000          2134 ns/op        1068 B/op         15 allocs/op
BenchmarkLogVanillaLevels-8         30000000            56.9 ns/op        64 B/op          2 allocs/op
BenchmarkLogScrubbing-8               100000         13231 ns/op        6597 B/op         38 allocs/op
BenchmarkLogScrubbingLevels-8       20000000            68.4 ns/op        16 B/op          1 allocs/op
BenchmarkLogScrubbingMulti-8          100000         15175 ns/op        7563 B/op         54 allocs/op
PASS
coverage: 26.3% of statements
```

After change:
```
goos: darwin
goarch: amd64
pkg: github.com/DataDog/datadog-agent/pkg/util/log
BenchmarkLogVanilla-8                1000000          2266 ns/op        1068 B/op         15 allocs/op
BenchmarkLogVanillaLevels-8         30000000            57.4 ns/op        64 B/op          2 allocs/op
BenchmarkLogScrubbing-8               100000         13462 ns/op        6680 B/op         39 allocs/op
BenchmarkLogScrubbingLevels-8       20000000            67.5 ns/op        16 B/op          1 allocs/op
BenchmarkLogScrubbingMulti-8          100000         15106 ns/op        7644 B/op         55 allocs/op
PASS
coverage: 27.8% of statements
```