@KSerrania 2019-05-28
@truthbk A possible explanation for why some benchmarks went down is that I removed a thing that was strange: after applying the replacers to a line (which is a `[]byte`, we used to cast this line to a `string` and append it to a result string. And then when the entire file is taken care of, we re-cast this `string` into a `[]byte`.
Since there was no need to do these two casts (and that I needed a `[]byte` at the end of the replacers loop to apply my new replacer), I removed the intermediary `string` and kept everything in a `[]byte`.